Move RoomSubscriptionSettings to protocol directory
