+++ packages/liveblocks-core/src/api-client.ts
@@ -55,7 +55,7 @@ import type {
 import type { HistoryVersion } from "./protocol/VersionHistory";
 import type { TextEditorType } from "./types/Others";
 import type { Patchable } from "./types/Patchable";
-import type { RoomSubscriptionSettings } from "./types/RoomSubscriptionSettings";
+import type { RoomSubscriptionSettings } from "./protocol/RoomSubscriptionSettings";
 import { PKG_VERSION } from "./version";
 
 export interface RoomHttpApi<M extends BaseMetadata> {
