+++ packages/liveblocks-core/src/room.ts
@@ -98,7 +98,7 @@ import type {
   TextEditorType,
 } from "./types/Others";
 import type { Patchable } from "./types/Patchable";
-import type { RoomSubscriptionSettings } from "./types/RoomSubscriptionSettings";
+import type { RoomSubscriptionSettings } from "./protocol/RoomSubscriptionSettings";
 import type { User } from "./types/User";
 import { PKG_VERSION } from "./version";
